Commit checkpoint for adding in preparation.protonate.py file
